Revert "[aot_inductor] move CppWrapperCodeGen into a separate file (#119491)"
This reverts commit 760056bbdc552314e7e81adc45e11766ac0f333c.
