+++ torch/_inductor/codegen/cpp.py
@@ -3370,7 +3370,7 @@ class CppScheduling(BaseScheduling):
         return tuple(tuple(map(V.graph.sizevars.simplify, s)) for s in sizes)
 
     def get_kernel_group(self):
-        from .cpp_wrapper_cpu import CppWrapperCodeGen
+        from .wrapper import CppWrapperCodeGen
 
         self.kernel_group: Union[CppWrapperKernelGroup, KernelGroup]
         if isinstance(V.graph.wrapper_code, CppWrapperCodeGen):
