+++ torch/_inductor/codegen/cpp_wrapper_cuda.py
@@ -11,8 +11,7 @@ from .. import config
 from ..codecache import CudaKernelParamCache
 from ..triton_heuristics import grid as default_grid
 from ..virtualized import V
-from .cpp_wrapper_cpu import CppWrapperCodeGen
-from .wrapper import SymbolicCallArg
+from .wrapper import CppWrapperCodeGen, SymbolicCallArg
 
 
 def is_int(s: str) -> bool:
